restructuring code - partially working
