+++ python/whylogs/api/logger/experimental/logger/actor/process_rolling_logger_messages.py
@@ -99,7 +99,7 @@ class ProcessLoggerStatus:
 
 
 @dataclass
-class ProcessLoggerStatusMessage:
+class ProcessStatusMessage:
     id: str = field(default_factory=lambda: str(uuid.uuid4()))
     timeout: Optional[float] = None
